pyproject.toml
[tool.poetry]
name = "bigquery-utils"
version = "0.1.0"
description = ""
authors = ["PeDiot <p.emmanuel.diot@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-cloud-bigquery = "^3.13.0"
pandas = "^2.1.3"
ipython = "^8.17.2"
db-dtypes = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
